+++ packages/design-tokens/web/properties/colors.v2.json5
