+++ packages/design-tokens/web/properties/shadows.v2.json5
