+++ packages/tokens-builder/configs/css.js
@@ -2,87 +2,7 @@ const fs = require('fs');
 const path = require('node:path');
 
 const paletteColors = ['black', 'blue', 'cyan', 'green', 'grey', 'orange', 'purple', 'red', 'white', 'yellow'];
-const newPaletteColors = [
-    'blue',
-    'slate',
-    'red',
-    'orange',
-    'yellow',
-    'green',
-    'teal',
-    'purple',
-    'darkBlue',
-    'darkSlate',
-    'darkRed',
-    'darkOrange',
-    'darkYellow',
-    'darkGreen',
-    'darkTeal',
-    'darkPurple',
-    'grey',
-    'darkGrey',
-    'greyA',
-    'darkGreyA',
-    'whiteA',
-    'blackA',
-    'blueA',
-    'slateA',
-    'redA',
-    'orangeA',
-    'yellowA',
-    'greenA',
-    'tealA',
-    'purpleA',
-    'darkBlueA',
-    'darkSlateA',
-    'darkRedA',
-    'darkOrangeA',
-    'darkYellowA',
-    'darkGreenA',
-    'darkTealA',
-    'darkPurpleA',
-    'yellowFixed',
-    'orangeFixed',
-    'darkYellowFixed',
-    'darkOrangeFixed',
-    'yellowFixedA',
-    'orangeFixedA',
-    'darkYellowFixedA',
-    'darkOrangeFixedA',
-    'white',
-    'black'
-];
 const semanticPaletteColors = ['theme', 'success', 'warning', 'error', 'contrast', 'white', 'black', 'purple'];
-const newSemanticPaletteColors = [
-    'contrast',
-    'contrastA',
-    'theme',
-    'themeA',
-    'error',
-    'errorA',
-    'warning',
-    'warningA',
-    'warningFixed',
-    'warningFixedA',
-    'success',
-    'successA',
-    'visited',
-    'visitedA',
-    'darkContrast',
-    'darkContrastA',
-    'darkTheme',
-    'darkThemeA',
-    'darkError',
-    'darkErrorA',
-    'darkWarning',
-    'darkWarningA',
-    'darkWarningFixed',
-    'darkWarningFixedA',
-    'darkSuccess',
-    'darkSuccessA',
-    'darkVisited',
-    'darkVisitedA'
-];
 
 // resolve components path if script used externally
 const componentsPath = fs.existsSync(path.join('node_modules', '@koobiq/design-tokens', 'web', 'components'))
@@ -131,31 +51,14 @@ const paletteByColorsConfig = paletteColors.map((color) => ({
     prefix: 'kbq'
 }));
 
-const newPaletteByColorsConfig = newPaletteColors.map((color) => ({
-    destination: `css/new/palette/${color}.css`,
-    format: 'kbq-css/variables',
-    filter: (token) => token.attributes.category === 'plt' && token.attributes.type === color,
-    prefix: 'kbq'
-}));
-
 const semanticPaletteConfig = semanticPaletteColors.map((color) => ({
     destination: `css/semantic-palette/${color}.css`,
     format: 'kbq-css/palette',
     filter: (token) =>
-        token.filePath.includes('colors.json5') &&
+        token.filePath.includes('colors.v1.json5') &&
         token.attributes.light &&
         token.attributes.type === color &&
         (token.attributes.palette || token.name.includes('default')),
-    prefix: 'kbq',
-    options: {
-        outputReferences: true
-    }
-}));
-
-const newSemanticPaletteConfig = newSemanticPaletteColors.map((color) => ({
-    destination: `css/new/semantic-palette/${color}.css`,
-    format: 'kbq-css/palette',
-    filter: (token) => token.attributes.category === 'semantic' && token.attributes.type === color,
     prefix: 'kbq',
     options: {
         outputReferences: true
@@ -167,26 +70,24 @@ module.exports = {
         transformGroup: 'kbq/css',
         files: [
             ...semanticPaletteConfig,
-            ...newSemanticPaletteConfig,
             {
                 destination: 'css/light/semantic-palette.css',
                 format: 'kbq-css/palette',
                 filter: (token) =>
                     token.attributes.light &&
-                    token.filePath.includes('colors.json5') &&
+                    token.filePath.includes('colors.v1.json5') &&
                     (token.attributes.palette || token.name.includes('default')),
                 prefix: 'kbq',
                 options: {
                     selector: '.kbq-light'
                 }
             },
-            // deprecated
             {
                 destination: 'css/dark/semantic-palette.css',
                 format: 'kbq-css/palette',
                 filter: (token) =>
                     token.attributes.dark &&
-                    token.filePath.includes('colors.json5') &&
+                    token.filePath.includes('colors.v1.json5') &&
                     (token.attributes.palette || token.name.includes('default')),
                 prefix: 'kbq',
                 options: {
@@ -196,14 +97,10 @@ module.exports = {
             {
                 destination: 'css/light/semantic-colors.css',
                 format: 'kbq-css/palette',
-                filter: (token) =>
-                    token.attributes.light &&
-                    token.filePath.includes('colors.json5') &&
-                    !semanticPaletteColors.includes(token.attributes.type),
+                filter: (token) => token.attributes.light && token.filePath.includes('colors.v1.json5'),
                 prefix: 'kbq',
                 options: {
-                    selector: '.kbq-light',
-                    outputReferences: true
+                    selector: '.kbq-light'
                 }
             },
             {
@@ -211,12 +108,11 @@ module.exports = {
                 format: 'kbq-css/palette',
                 filter: (token) =>
                     token.attributes.dark &&
-                    token.filePath.includes('colors.json5') &&
+                    token.filePath.includes('colors.v1.json5') &&
                     !semanticPaletteColors.includes(token.attributes.type),
                 prefix: 'kbq',
                 options: {
-                    selector: '.kbq-dark',
-                    outputReferences: true
+                    selector: '.kbq-dark'
                 }
             },
             {
@@ -241,27 +137,11 @@ module.exports = {
                 }
             },
             ...paletteByColorsConfig,
-            ...newPaletteByColorsConfig,
             {
                 destination: 'css/palette.css',
                 format: 'kbq-css/variables',
                 filter: (token) => token.attributes.category === 'palette',
                 prefix: 'kbq'
-            },
-            {
-                destination: 'css/new/palette.css',
-                format: 'kbq-css/variables',
-                filter: (token) => token.attributes.category === 'plt',
-                prefix: 'kbq'
-            },
-            {
-                destination: 'css/new/semantic-palette.css',
-                format: 'kbq-css/variables',
-                filter: (token) => token.attributes.category === 'semantic',
-                prefix: 'kbq',
-                options: {
-                    outputReferences: true
-                }
             },
             {
                 destination: 'css/light/shadows.css',
